requirements.txt
fastapi
uvicorn
feedparser
newspaper3k
spacy
sentence-transformers
faiss-cpu
chromadb
pydantic
python-docx
reportlab
pdfminer.six
transformers
torch
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
psycopg2-binary
redis
python-json-logger
slowapi
sentry-sdk
pytest
pytest-cov
httpx
